backend/requirements.txt
fastapi
uvicorn
langchain
chromadb
pypdf
sentence-transformers
openai
python-multipart
python-dotenv